Size the render targets from the window instead of a latched integer

+++ src/native_renderer.h
@@ -47,11 +47,14 @@ inline constexpr const char* kNativeRendererPluginName = "plume";
 // meaningful once the config files have been loaded.
 bool NativeRendererEnabled();
 
-// Supersampling factor the guest's 1280x720 frame is rendered at, 1..8. This is
-// the `resolution_scale` cvar, latched at boot; see NativeRenderScale's body for
-// why it cannot be a live read. Always 1 on the Xenos path, which owns the cvar
-// itself and does its own scaling.
-uint32_t NativeRenderScale();
+// Supersampling factor the guest's 1280x720 frame is rendered at, latched at
+// boot; see NativeRenderScale's body for why it cannot be a live read. Always 1
+// on the Xenos path, which owns the cvars itself and does its own scaling.
+//
+// Fractional, so it is not a texel-for-texel supersample any more: `render_scale`
+// gives it directly and `resolution_scale` remains the integer fallback the
+// Xenos backend shares.
+float NativeRenderScale();
 
 // Registers the cvars a GPU plugin would have registered, so selecting this
 // renderer does not silently lose them: `vsync` and `resolution_scale`, both of
